Renamed time.h->time_utils.h to prevent conflicts
